+++ Agent_TA/MarketOrder.py
@@ -3,22 +3,20 @@ This file is responsible for market order execution
 '''
 import pandas as pd
 import MetaTrader5 as mt5
-from connections import Connections
-
 
 class Market_Order_Execution():
     def __init__(self) -> None:
         
         self.order_result = None
     
-    def market_order(self, tick, symbol, volume, magic, stoploss, current_price, bagging_signal):
+    def market_order(self, tick, symbol, stoploss, take_prof, current_price, bagging_signal, volume=0.02, magic=100):
         '''
         This function will send the request to platform and request to open the position
         '''
         
         order_dict = {'Buy': 0, 'Sell': 1}
         price_dict = {'Buy': tick.ask, 'Sell': tick.bid}
-        
+
         # take_profit = {'buy': upper_band - stoploss, 'sell': lower_band + stoploss}
         stop_loss = {'Buy': current_price - stoploss, 'Sell': current_price + stoploss} 
         
@@ -28,9 +26,8 @@ class Market_Order_Execution():
             'volume': volume, 
             'type': order_dict[bagging_signal], # Buy or Sell
             'price': price_dict[bagging_signal], # Buy at Curent ask price
-            #'deviation': deviation, #INT # Within deviation range when market open THEN make the DEAL
             'magic': magic, #INT # Each order have a unique magic number to identify 
-            #'tp': take_profit[order_type], # Take Profit
+            'tp': take_prof, # Take Profit
             'sl': stop_loss[bagging_signal], # Stop Loss default
             'comment': 'python order',
             'type_time': mt5.ORDER_TIME_GTC, # Good till Cancel order
@@ -41,8 +38,5 @@ class Market_Order_Execution():
         
         print(order_result)
                     
-        # Store Results of Orders
-        Connections().sql_connect(table_db_name = 'market_orders', table=pd.DataFrame([order_result]), ifexists='append')
-
         self.order_result = order_result
