+++ Agent_TA/run.py
@@ -1,9 +1,9 @@
 # Python Libraries
+import time
 import pandas as pd
 from Utils import Utils
 import MetaTrader5 as mt5
-from Decision import Decision
-from connections import Connections
+import Config as config
 from RiskManager import Risk_Manager
 from TrendMetrics import Trend_Metrics
 from CandlesPatterns import Candles_Patterns
@@ -23,21 +23,15 @@ class engine():
 
     def main(self, ut, auto_trade: bool, full_list: bool, ticker, params):
         
-        # Risk Manager
-        risk = Risk_Manager()
         # Trend Metrics
         tm = Trend_Metrics()
-        tm.get_crossover(ut.data, params.BEST_FIRST_PERIOD_Ema_Cross, params.BEST_SECOND_PERIOD_Ema_Cross, params.BEST_THIRD_PERIOD_Ema_Cross, type = 1)
-        tm.get_sma_bands(ut.data, params.BEST_LENGHT_Bbands ,params.BEST_STD_Bbands)
-        tm.get_rsi(ut.data, ut.RSI_PERIOD, ut.OVERBOUGHT_LEVEL, ut.OVERSOLD_LEVEL)
-        tm.get_volume_pvi_nvi(ut.data)
-        tm.get_macd(ut.data)
-        tm.combined_rsi_bollinger(ut.data, ut.SMA_PERIOD ,ut.STANDARD_DEVIATION_SMA, ut.RSI_PERIOD)
+        tm.get_crossover(ut.data, params.Best_Ema1_Cross, params.Best_Ema2_Cross, params.Best_Ema3_Cross)
+        tm.get_sma_bands(ut.data, params.Best_Ema1_Bbands ,params.Best_std_Bbands)
+        tm.get_rsi(ut.data, params.Best_Ema1_Bbands, params.Rsi_overbought, params.Rsi_oversold)
+        tm.get_volume_profile(ut.data_1h)
         
         # Instantiate the Candles_Patterns class
         cp = Candles_Patterns()
-
-        # Call the methods of Candles_Patterns class
         cp.candle_basic_signal(ut.data)
         cp.engulfing(ut.data)
         cp.morning_star(ut.data)
@@ -45,52 +39,59 @@ class engine():
         cp.hammer(ut.data)
         cp.hanging_man(ut.data)
 
-    
+        # Risk Manager
+        risk = Risk_Manager()        
+        stloss = risk.risk_manager(cp.result_df)
+
         # Bagging Signal
-        bg = Decision(ticker)
-        
-        decision_signal = bg.get_decision(tm.result_df, cp.result_df)
+        decision_signal = risk.get_decision(tm.result_df, cp.result_df)
 
         # ==== Results Report ==== #
-        # Save dos resultados
-
-        # Display the result dataframe
-        print('Basic Information/Settings\n\n\nTrend Metrics\n{}\n\nCandle Patterns\n{}\n\nDecision\n{}\n\n'.format(tm.result_df, cp.result_df, bg.decision_df))
+        print('\033[91mBasic Information/Settings:\033[0m\n\n   \
+              \033[91mTrend Metrics:\033[0m\n{}\n\n \
+              Support: {}\n \
+              Resistence: {}\n\n \
+              \033[91mCandle Patterns:\033[0m\n{}\n\n  \
+              \033[91mDecision:\033[0m\n{}\n\n  \
+              \033[91mCriteria:\033[0m\n{}\n\n' \
+              .format(tm.result_df, tm.support, tm.resistence, cp.result_df, decision_signal, params))
+        
+        time.sleep(2)
 
         if auto_trade and not full_list:        
-            # Utils
+
             ut.get_exposure(ticker)
-            tls = Tailing_Stoploss()
-            tls.modify_position_tailling(risk.TAILLING_VALUE)
-            
-            if not mt5.positions_total() and decision_signal != 'Flat':
-                
-                Lot_Size = risk.market_order_size(params.WinRate_avg, ut.position_current_price)
-
-                mk = Market_Order_Execution()
-                mk.market_order(ut.tick, ticker, Lot_Size, ut.MAGIC, risk.DEFAULT_STOP_LOSS, ut.position_current_price, decision_signal)
-                
-                # Store Results of Metrics
-                metrics_results = pd.concat([tm.result_df, cp.result_df], axis=0)
-                # Connections().sql_connect(table_db_name = 'metrics_results', table=metrics_results, ifexists='append')
-
+            if stloss == "tail":
+                tls = Tailing_Stoploss()
+                tls.modify_position_tailling(sup=tm.support, resis=tm.resistence, tailling_value = 0.01)
+                if not mt5.positions_total() and decision_signal == 'Sell':
+                    mk = Market_Order_Execution()
+                    mk.market_order(ut.tick, ticker, 0.01, tm.support, ut.position_current_price, decision_signal, ut.VOLUME, ut.MAGIC)
+                if not mt5.positions_total() and decision_signal == 'Buy':
+                    mk = Market_Order_Execution()
+                    mk.market_order(ut.tick, ticker, 0.01, tm.resistence, ut.position_current_price, decision_signal, ut.VOLUME, ut.MAGIC)
+            elif stloss != "tail":
+                if not mt5.positions_total() and decision_signal == 'Sell':
+                    mk = Market_Order_Execution()
+                    mk.market_order(ut.tick, ticker, stloss, tm.support, ut.position_current_price, decision_signal, ut.VOLUME, ut.MAGIC)
+                if not mt5.positions_total() and decision_signal == 'Buy':
+                    mk = Market_Order_Execution()
+                    mk.market_order(ut.tick, ticker, stloss, tm.resistence, ut.position_current_price, decision_signal, ut.VOLUME, ut.MAGIC)
+        
         elif auto_trade and full_list:
 
             ut.get_exposure(ticker)
-            tls = Tailing_Stoploss()
-            tls.modify_position_tailling(risk.TAILLING_VALUE)
-
-            if mt5.positions_get(symbol=ticker) == () and decision_signal != 'Flat':
-
-                Lot_Size = risk.market_order_size(params.WinRate_avg, ut.position_current_price)
-
-                mk = Market_Order_Execution()
-                mk.market_order(ut.tick, ticker, Lot_Size, ut.MAGIC, risk.DEFAULT_STOP_LOSS, ut.position_current_price, decision_signal)
-                
-                # Store Results of Metrics
-                metrics_results = pd.concat([tm.result_df, cp.result_df], axis=0)
-                # Connections().sql_connect(table_db_name = 'metrics_results', table=metrics_results, ifexists='append')
-
+            if stloss == "tail":
+                tls = Tailing_Stoploss()
+                tls.modify_position_tailling(sup=tm.support, resis=tm.resistence, tailling_value = 0.01)
+                if mt5.positions_get(symbol=ticker) == () and decision_signal != 'Flat':
+                    mk = Market_Order_Execution()
+                    mk.market_order(ut.tick, ticker, 0.01, ut.position_current_price, decision_signal, ut.VOLUME, ut.MAGIC)
+            elif stloss != "tail":
+                if mt5.positions_get(symbol=ticker) == () and decision_signal != 'Flat':
+                    mk = Market_Order_Execution()
+                    mk.market_order(ut.tick, ticker, stloss, ut.position_current_price, decision_signal, ut.VOLUME, ut.MAGIC)
+        
     def execute_strategy(self, market_type, timeframe, ticker, autotrade_set, stop):
         '''
         This function represents the execution selection according to user request 
@@ -98,8 +99,7 @@ class engine():
         
         if market_type == 'Cambial':
             Utils.login()
-        
-        
+                
             if ticker == 'Top10':
                 if autotrade_set == 'on': 
                     while stop == False:
@@ -110,9 +110,9 @@ class engine():
                                     ) 
                         
                         tickers_to_process = tickers['Ticker'].head(10).tolist()
-                        best_params_top10 = Params_Optimization(tickers_to_process, market_type, timeframe).best_params   
+                        best_params_top = Params_Optimization(tickers_to_process, market_type, timeframe).best_params   
 
-                        for _, row in best_params_top10.iterrows():
+                        for _, row in best_params_top.iterrows():
                             utils_instance = Utils(row['Ticker'], market_type, timeframe)
                             self.main(ut=utils_instance, auto_trade=True, full_list=True, ticker=row['Ticker'], params=row)
                 
@@ -124,10 +124,10 @@ class engine():
                                     columns=['Ticker', 'Description', 'Currency']
                                     ) 
                         
-                        tickers_to_process = tickers['Ticker'].head(10).tolist()
-                        best_params_top10 = Params_Optimization(tickers_to_process, market_type, timeframe).best_params   
+                        tickers_to_process = tickers['Ticker'].head(4).tolist()
+                        best_params_top = Params_Optimization(tickers_to_process, market_type, timeframe).best_params   
 
-                        for _, row in best_params_top10.iterrows():
+                        for _, row in best_params_top.iterrows():
                             utils_instance = Utils(row['Ticker'], market_type, timeframe)
                             self.main(ut=utils_instance, auto_trade=False, full_list=False, ticker=row['Ticker'], params=row)
             
@@ -136,9 +136,9 @@ class engine():
                 crypto = crypto[crypto['Path'].str.contains("crypto", case=False)]
                 tickers_to_process = crypto['Ticker'].tolist()
 
-                best_params_top10 = Params_Optimization(tickers_to_process, market_type, timeframe).best_params   
+                best_params_top = Params_Optimization(tickers_to_process, market_type, timeframe).best_params  
 
-                for _, row in best_params_top10.iterrows():
+                for _, row in best_params_top.iterrows():
                     utils_instance = Utils(row['Ticker'], market_type, timeframe)
                     self.main(ut=utils_instance, auto_trade=False, full_list=False, ticker=row['Ticker'], params=row)
     
@@ -158,9 +158,3 @@ class engine():
                         for _, row in best_params.iterrows():
                             utils_instance = Utils(row['Ticker'], market_type, timeframe)
                             self.main(ut=utils_instance, auto_trade=False, full_list=False, ticker=row['Ticker'], params=row)
-
-
-
-    # cd C:\Users\joaom\Matris\TradingRobot\Pyrobot\Application\Agent_TA   python Main.py
-
-    # Notes
